fix(script/install_debian): test if VScode is already here (#4)

+++ scripts/install_debian.sh
@@ -1,8 +1,13 @@
 #! /bin/bash
 
-wget -O code.deb https://go.microsoft.com/fwlink/?LinkID=760868
-sudo apt install -y git curl python3 python3-pip ./code.deb
-rm code.deb
+sudo apt install -y git curl python3 python3-pip 
+# The following test is needed in case VScode was installed by other
+# means (e.g. using Ubuntu snap)
+if ! which code; then
+  wget -O code.deb https://go.microsoft.com/fwlink/?LinkID=760868
+  sudo apt install -y ./code.deb
+  rm code.deb
+fi
 code --install-extension jroesch.lean
 wget https://raw.githubusercontent.com/Kha/elan/master/elan-init.sh
 bash elan-init.sh -y
